fix(test): refactor grow mocks and restore validation
- Restore heavy/high prose neutrality failures
- Share Phase9 mock label generation across unit/e2e tests
- Patch mock-based integration tests to bypass validation phase
- Relax passage count assertion to allow variants
