+++ tests/unit/test_grow_algorithms.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import re
 from typing import TYPE_CHECKING, Any
 from unittest.mock import AsyncMock, MagicMock
 
@@ -26,7 +25,6 @@ from questfoundry.graph.mutations import GrowErrorCategory
 from questfoundry.models.grow import (
     Arc,
     AtmosphericDetail,
-    ChoiceLabel,
     PathMiniArc,
     Phase3Output,
     Phase4aOutput,
@@ -52,6 +50,7 @@ from tests.fixtures.grow_fixtures import (
     make_single_dilemma_graph,
     make_two_dilemma_graph,
 )
+from tests.fixtures.grow_mock_llm import build_phase9_output
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -3406,113 +3405,8 @@ def _make_grow_mock_model(graph: Graph, project_path: Path | None = None) -> Mag
     phase9b_output = Phase9bOutput(proposals=[])
     phase9c_output = Phase9cOutput(hubs=[])
 
-    def _extract_transition_pairs(messages: list[Any]) -> list[tuple[str, str]]:
-        text = "\n".join(
-            getattr(message, "content", str(message)) for message in messages if message is not None
-        )
-        text = text.split("Semantic validation errors", 1)[0]
-        pairs: list[tuple[str, str]] = []
-        is_divergence = "Divergence at" in text or "Divergence Points to Label" in text
-
-        if is_divergence:
-            # Fallback: parse divergence blocks
-            current_from: str | None = None
-            for line in text.splitlines():
-                from_match = re.search(r"Divergence at (passage::[^\s,:]+)", line)
-                if from_match:
-                    current_from = from_match.group(1)
-                    continue
-                to_match = re.search(r"-\s*(passage::[^\s,:]+)", line)
-                if current_from and to_match:
-                    pair = (current_from, to_match.group(1))
-                    if pair not in pairs:
-                        pairs.append(pair)
-        else:
-            for line in text.splitlines():
-                if "→" not in line and "->" not in line:
-                    continue
-                ids = re.findall(r"passage::[^\s,:]+", line)
-                if len(ids) >= 2:
-                    pair = (ids[0], ids[1])
-                    if pair not in pairs:
-                        pairs.append(pair)
-
-            # Fallback: parse valid_from_ids/valid_to_ids and zip (continue labels)
-            def _extract_ids(segment: str) -> list[str]:
-                tokens = re.split(r"[\s,]+", segment)
-                ids: list[str] = []
-                for token in tokens:
-                    cleaned = token.strip().strip('"').strip("'").strip(")").strip("]").rstrip(":")
-                    if cleaned.startswith("passage::"):
-                        ids.append(cleaned)
-                return ids
-
-            from_ids: list[str] = []
-            to_ids: list[str] = []
-            if "valid_from_ids:" in text:
-                segment = text.split("valid_from_ids:", 1)[1].split("valid_to_ids:", 1)[0]
-                from_ids = _extract_ids(segment)
-            if "valid_to_ids:" in text:
-                segment = text.split("valid_to_ids:", 1)[1]
-                segment = segment.split("output_language_instruction", 1)[0]
-                to_ids = _extract_ids(segment)
-            if from_ids and to_ids and len(from_ids) == len(to_ids):
-                for src, dst in zip(from_ids, to_ids, strict=True):
-                    pair = (src, dst)
-                    if pair not in pairs:
-                        pairs.append(pair)
-
-        if pairs:
-            return pairs
-
-        # Last resort: extract first two passage IDs from any line
-        for line in text.splitlines():
-            ids = re.findall(r"passage::[^\s,:]+", line)
-            if len(ids) >= 2:
-                pair = (ids[0], ids[1])
-                if pair not in pairs:
-                    pairs.append(pair)
-        return pairs
-
     def _build_phase9_output(messages: list[Any]) -> Phase9Output:
-        text = "\n".join(
-            getattr(message, "content", str(message)) for message in messages if message is not None
-        )
-        is_divergence_prompt = "Divergence Points to Label" in text
-
-        if project_path is not None:
-            from questfoundry.graph.grow_algorithms import find_passage_successors
-
-            snapshot = Graph.load(project_path / "graph.db")
-            successors = find_passage_successors(snapshot)
-            pairs: list[tuple[str, str]] = []
-            if successors:
-                if is_divergence_prompt:
-                    pairs = [
-                        (p_id, succ.to_passage)
-                        for p_id, succ_list in successors.items()
-                        if len(succ_list) > 1
-                        for succ in succ_list
-                    ]
-                else:
-                    pairs = [
-                        (p_id, succ_list[0].to_passage)
-                        for p_id, succ_list in successors.items()
-                        if len(succ_list) == 1
-                    ]
-
-            if pairs:
-                labels = [
-                    ChoiceLabel(from_passage=src, to_passage=dst, label="continue")
-                    for src, dst in pairs
-                ]
-                return Phase9Output(labels=labels)
-
-        labels = [
-            ChoiceLabel(from_passage=src, to_passage=dst, label="continue")
-            for src, dst in _extract_transition_pairs(messages)
-        ]
-        return Phase9Output(labels=labels)
+        return build_phase9_output(messages, project_path=project_path)
 
     # Phase 4e: PathMiniArc is called per-path (single object, not wrapper)
     # The mock will return a generic PathMiniArc for any path
@@ -3562,6 +3456,8 @@ def _make_grow_mock_model(graph: Graph, project_path: Path | None = None) -> Mag
 class TestPhaseIntegrationEndToEnd:
     @staticmethod
     def _patch_validation(monkeypatch: pytest.MonkeyPatch) -> None:
+        # The mock LLM does not guarantee semantic validity; bypass validation
+        # so these tests exercise phase wiring rather than validation outcomes.
         from questfoundry.graph import grow_validation as grow_validation
         from questfoundry.graph.grow_validation import ValidationCheck, ValidationReport
